Rename controller common processes
