+++ backend/apps/controller/common/models.py
@@ -1,7 +1,7 @@
-from common.system_processes import database_recover
+from common.processes import database_recover
 from flask_migrate import Migrate
 from flask_sqlalchemy import SQLAlchemy
-from resources.errors import print_message
+from common.errors import print_message
 from sqlalchemy.sql import expression
 import bcrypt
