+++ backend/apps/controller/resources/app_store_routes.py
@@ -3,7 +3,7 @@ from common.models import App_Store
 from flask_jwt_extended import jwt_required
 from flask_restful import abort
 from flask_restful import Resource
-from resources.errors import print_message
+from common.errors import print_message
 import json
 import os
 import requests
