+++ backend/apps/controller/resources/dev_routes.py
@@ -1,5 +1,5 @@
 from common.models import User
-from common.system_processes import check_internet
+from common.processes import check_internet
 from flask import request
 from flask_jwt_extended import jwt_required
 from flask_restful import Resource
