+++ backend/apps/controller/resources/download_routes.py
@@ -2,7 +2,7 @@ from flask import request
 from flask import Response
 from flask_jwt_extended import jwt_required
 from flask_restful import Resource
-from resources.errors import print_message
+from common.errors import print_message
 import json
 import shutil
 import time
