+++ backend/apps/controller/resources/supervisor_routes.py
@@ -1,4 +1,4 @@
-from common.system_processes import curl
+from common.processes import curl
 from flask import request
 from flask_jwt_extended import jwt_required
 from flask_restful import Resource
