+++ backend/apps/controller/resources/wifi_routes.py
@@ -1,11 +1,11 @@
 from common.models import User
-from common.system_processes import check_internet
+from common.processes import check_internet
 from common.wifi import wifi
 from flask import request
 from flask_jwt_extended import jwt_required
 from flask_restful import abort
 from flask_restful import Resource
-from resources.errors import print_message
+from common.errors import print_message
 import threading
