+++ backend/apps/controller/run.py
@@ -1,6 +1,6 @@
 from common.models import db
 from common.models import migrate
-from common.system_processes import curl
+from common.processes import curl
 from dotenv import load_dotenv
 from flask import Flask
 from flask_cors import CORS
@@ -14,8 +14,8 @@ from resources.docker_routes import docker_run
 from resources.download_routes import download_fetch
 from resources.download_routes import download_stop
 from resources.download_routes import download_stream
-from resources.errors import errors
-from resources.errors import print_message
+from common.errors import errors
+from common.errors import print_message
 from resources.filemanager_routes import filemanager_copy
 from resources.filemanager_routes import filemanager_delete
 from resources.filemanager_routes import filemanager_file_size
